requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
decorator==4.4.2
Django==5.0.2
django-celery-beat==2.6.0
django-timezone-field==6.1.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
instagrapi==2.1.1
kombu==5.3.5
moviepy==1.0.3
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pluggy==1.4.0
proglog==0.1.10
prompt-toolkit==3.0.43
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PySocks==1.7.1
pytest==8.1.1
pytest-django==4.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.2
requests==2.31.0
selenium==4.9.0
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
wsproto==1.2.0
